Converted remaining ingredients tests to use entityTestFactory
