+++ api/src/controllers/miscs.test.js
@@ -1,4 +1,4 @@
-import { randomString, randomFloat, getRandomArrayMembers, randomInt } from "../utils/helpers.js";
+import { randomString, randomFloat, getRandomArrayMembers } from "../utils/helpers.js";
 import entityTestFactory from "../../test/entityTestFactory.js";
 import {MISC_TYPES} from "../services/misc.js";
